// src/priority/types/priority.types.ts
export type PriorityStatus = 'OPE' | 'CLO' | 'CAN';

export type MonthlyClass =
  | 'ABIERTAS'
  | 'NO_CUMPLIDAS_ATRASADAS_DEL_MES'
  | 'ANULADAS'
  | 'CUMPLIDAS_A_TIEMPO'
  | 'CUMPLIDAS_ATRASADAS_DEL_MES'
  | 'CUMPLIDAS_ATRASADAS_MESES_ANTERIORES'
  | 'CUMPLIDAS_DE_OTRO_MES'
  | 'NO_CUMPLIDAS_ATRASADAS_MESES_ANTERIORES';